FIX-2269 move breadcrumbs into own file
